+++ packages/gamut-styles/src/variables/borderRadii.ts
@@ -1,0 +1,8 @@
+export const borderRadii = {
+  none: '0px',
+  sm: '2px',
+  md: '4px',
+  lg: '8px',
+  xl: '16px',
+  full: '999px',
+};