+++ packages/gamut/src/Tip/shared/styles.tsx
@@ -181,7 +181,7 @@ export const toolTipBodyCss = {
   color: 'text',
   border: 1,
   boxShadow: 'none',
-  borderRadius: '3px',
+  borderRadius: 'sm',
   display: 'inline-block',
   fontSize: 14,
   lineHeight: 'base',